feat(repository): change type List

+++ src/main/java/dev/pdrotmz/EncurtadorURL/Repository/LinkRepository.java
@@ -3,7 +3,7 @@ package dev.pdrotmz.EncurtadorURL.Repository;
 import dev.pdrotmz.EncurtadorURL.Model.Link;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LinkRepository extends JpaRepository<Link, Integer> {
+public interface LinkRepository extends JpaRepository<Link, Long> {
 
     // Query Personalizada
     Link findByUrlOriginal(String urlEncurtada);
